# 2) Optimize byte pair merge for small and big character sequences - 8.2s to 3.9s