Continuing https://github.com/knuddelsgmbh/jtokkit/pull/75 - note that the first few commits are repeated here, will be eliminated by rebase once the other one's merged.

The original byte pair merge algorithm diverges quickly for longer character sequences in a superlinear way - e.g. a 20000 character word (e.g. 2500 tokens) can take several seconds to tokenize.
<img width="500" src="https://github.com/knuddelsgmbh/jtokkit/assets/1841944/188ec58c-2061-473c-8a1f-e787b5fb2d4a">

For bigger character sequences we're switching to a linear(ithmic) algorithm at around 500 characters (below which the current one is faster):
<img width="500" src="https://github.com/knuddelsgmbh/jtokkit/assets/1841944/ad95e474-f788-4b82-b0aa-06312b83db26">

The change also includes an optimization for just token counting - when the tokens themselves aren't important.

Before (i.e. assuming https://github.com/knuddelsgmbh/jtokkit/pull/75 was merged):
```
Benchmark                                              (dataFolderPath)  Mode  Cnt  Score   Error  Units
SingleThreadedBenchmark.benchmarkCl100kBase                        data    ss   10  8.273 ± 0.116   s/op
SingleThreadedBenchmark.benchmarkP50kBase                          data    ss   10  9.762 ± 0.101   s/op
SingleThreadedBenchmark.benchmarkP50kEdit                          data    ss   10  9.790 ± 0.056   s/op
SingleThreadedBenchmark.benchmarkR50kBase                          data    ss   10  8.750 ± 0.065   s/op
```

After:
```
Benchmark                                              (dataFolderPath)  Mode  Cnt  Score   Error  Units
SingleThreadedBenchmark.benchmarkCl100kBase                        data    ss   10  4.547 ± 0.056   s/op
SingleThreadedBenchmark.benchmarkCl100kBaseTokenCount              data    ss   10  3.944 ± 0.031   s/op
SingleThreadedBenchmark.benchmarkP50kBase                          data    ss   10  5.427 ± 0.065   s/op
SingleThreadedBenchmark.benchmarkP50kEdit                          data    ss   10  5.375 ± 0.062   s/op
SingleThreadedBenchmark.benchmarkR50kBase                          data    ss   10  5.073 ± 0.063   s/op
```

Please review [commit-by-commit](https://github.com/knuddelsgmbh/jtokkit/pull/76/commits) for the changes to make sense:
<img width="500" alt="image" src="https://github.com/knuddelsgmbh/jtokkit/assets/1841944/ca3ae0ce-6863-430e-83f4-85d6848ef40b">